symbolic setup for views, reshape, transpose
